refactor: fix core coordinator dependency direction
Move all device-domain-pure modules from coordinator/ to core/ so that
core/ has no imports from coordinator/. The target dependency direction
is now enforced: core/ → transport/registers/modbus/scanner, and
coordinator/ → core/.

Moved to core/:
- connection_lifecycle, connection_state, connection_test, disconnect,
  transport_select (connection helpers)
- register_groups, register_processing, runtime_state (register helpers)
- scanner_kwargs (scanner construction)
- retry, read_common, read_batches, read_bits, runtime_io (IO helpers)

Extracted to core/ (split from coordinator files):
- core/scan_helpers.py: normalise_cached_register_name,
  normalise_available_registers (from coordinator/scan.py)
- core/write_path.py: encode_write_value, SingleWritePlan
  (from coordinator/write_path.py)

CoordinatorConfig moved from coordinator/models.py to core/models.py
(already present; coordinator/models.py deleted).

Updated coordinator/ to import from core/, updated _setup.py to use
coordinator_config_from_entry directly, fixed async_write_register in
core/client_registers.py, updated all test imports to canonical paths,
and added tests/test_dependency_direction.py enforcement test.
