+++ custom_components/thessla_green_modbus/coordinator/scan.py
@@ -3,10 +3,13 @@
 from __future__ import annotations
 
 import logging
-import re
 from typing import Any
 
 from ..const import DEFAULT_NAME, KNOWN_MISSING_REGISTERS, UNKNOWN_MODEL
+from ..core.scan_helpers import (  # noqa: F401 – re-exported for coordinator internal use
+    normalise_available_registers,
+    normalise_cached_register_name,
+)
 from ..scanner.device_info import DeviceCapabilities
 
 _LOGGER = logging.getLogger(__name__)
@@ -45,26 +48,6 @@ def load_full_register_list(coordinator: Any) -> None:
         "Loaded full register list: %d total registers",
         sum(len(regs) for regs in coordinator.available_registers.values()),
     )
-
-
-def normalise_cached_register_name(name: str) -> str:
-    """Normalise cached register names to current canonical form."""
-    match = re.fullmatch(r"([es])(\d+)", name)
-    if match:
-        return f"{match.group(1)}_{int(match.group(2))}"
-    return name
-
-
-def normalise_available_registers(
-    coordinator: Any, available: dict[str, list[str] | set[str]]
-) -> dict[str, set[str]]:
-    """Return available register names in canonical form."""
-    normalised: dict[str, set[str]] = {}
-    for reg_type, names in available.items():
-        if not isinstance(names, list | set):
-            continue
-        normalised[reg_type] = {normalise_cached_register_name(str(name)) for name in names}
-    return normalised
 
 
 def firmware_lacks_known_missing(firmware: Any) -> bool:
