+++ custom_components/thessla_green_modbus/core/client.py
@@ -22,7 +22,6 @@ from ..const import (
     holding_registers,
     input_registers,
 )
-from ..coordinator.models import CoordinatorConfig
 from ..scanner import DeviceCapabilities
 from ..transport.base import BaseModbusTransport
 from ..utils import utcnow as _utcnow
@@ -31,6 +30,7 @@ from .client_connection import _DeviceClientConnectionMixin
 from .client_registers import _DeviceClientRegistersMixin
 from .client_scanner import _DeviceClientScannerMixin
 from .io_mixin import _ModbusIOMixin
+from .models import CoordinatorConfig
 
 if TYPE_CHECKING:
     from homeassistant.config_entries import ConfigEntry
