+++ custom_components/thessla_green_modbus/core/connection_lifecycle.py
