+++ custom_components/thessla_green_modbus/core/read_common.py
