+++ custom_components/thessla_green_modbus/core/register_groups.py
