+++ custom_components/thessla_green_modbus/core/scanner_kwargs.py
