+++ tests/test_coordinator_errors.py
@@ -4,7 +4,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 from custom_components.thessla_green_modbus.coordinator import ThesslaGreenModbusCoordinator
-from custom_components.thessla_green_modbus.coordinator.retry import _PermanentModbusError
+from custom_components.thessla_green_modbus.core.retry import _PermanentModbusError
 
 
 @pytest.mark.asyncio
