+++ tests/test_coordinator_io.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, MagicMock
 
 import pytest
 from custom_components.thessla_green_modbus.coordinator import ThesslaGreenModbusCoordinator
-from custom_components.thessla_green_modbus.coordinator.retry import _PermanentModbusError
+from custom_components.thessla_green_modbus.core.retry import _PermanentModbusError
 from pymodbus.exceptions import ModbusIOException
