+++ tests/test_coordinator_scan_cache_extra.py
@@ -1,4 +1,4 @@
-from custom_components.thessla_green_modbus.coordinator.register_processing import (
+from custom_components.thessla_green_modbus.core.register_processing import (
     create_consecutive_groups,
 )
