+++ tests/test_optimized_integration_updates.py
@@ -104,7 +104,7 @@ class TestThesslaGreenModbusCoordinator:
         )
 
     def test_register_grouping(self, coordinator_data):
-        from custom_components.thessla_green_modbus.coordinator.register_processing import (
+        from custom_components.thessla_green_modbus.core.register_processing import (
             create_consecutive_groups,
         )
 
@@ -200,7 +200,7 @@ class TestThesslaGreenDeviceScanner:
 class TestPerformanceOptimizations:
     @pytest.mark.asyncio
     async def test_register_grouping_performance(self):
-        from custom_components.thessla_green_modbus.coordinator.register_processing import (
+        from custom_components.thessla_green_modbus.core.register_processing import (
             create_consecutive_groups,
         )
